BarCode-Scanner/railfit_scanner/requirements.txt
Flask==2.3.3
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
Werkzeug==2.3.7
scipy==1.11.3
